API: Add Attach call to Rancher storage API
Call `attach` inside `docker.volume.mount` can result in docker daemon hang.
We've experienced this in EBS driver. The `attach` API here will be used by
rancher-agent to setup the volume before call into `docker.volume.mount`.

In the future we may distingiush so called `in-band` and `out-of-band`
operations of `attach`. If driver is `out-of-band`, we won't call into `attach`
during `docker.volume.mount` and would assume the device is already prepared
and ready to be used. It will fit the driver that would involve container
operations during `attach` call, which can result in Docker deadlock.

[sheng: commit message, remove detach and in-band logic and device list and
other fixes]

Signed-off-by: Sheng Yang <sheng@yasker.org>