rewrite garbage collection and fix ebs
we use a map of mountpoint to track if the mountpoint is still active. If a mountpoint is not used by any of the container, we garbage collect it and unmount the path. At startup we build an
initial map based on docker inspect. Then we watch the container events to make sure: 1. every container start will log an entry on the map. 2. every container destory will delete the entry in the map.